Here's the Battleship source code I wrote. It's split into 3 parts: 1. Main - main.cpp 2. headers - Player.h and Ship.h 3. source code - Player.cpp and Ship.cpp
This is a full, playable 2-player Battleship game. Create a netbeans project, and
include these files. Run it, let me know what you think. We can either build
on this, and concentrate on hooking it up to a GUI frontend; We can also rip
this apart, and just use it as a framework to create our own classes, depending
on what else we want to implement(i.e., AI for a computer opponent.) However, we
DO NOT HAVE TO use my code. We can use some other source code and hook that
up into a GUI. I figured this could be a time saver, and just port this to
our project. Let me know what you guys think.. Thanks!

-Ezra G.-

Signed-off-by: Tekktronic <ezgmusic@gmail.com>