Oscar- easy and medium functions
Hey guys, here are the easy and medium mode functions I've been working
on. They are by no means the final version but I think it is a good
start. I have yet to implement them with Ezra's classes but hopefully it
will be easier once i all the logic down. Like I said they are NOT the
final versions and there are a couple of bugs that i need to take care
of.

Give it a run and let me know what you think.